assets/code/5-algorithm-problems-to-get-your-brain-working/requirements.txt
igraph==1.0.0
numpy==2.3.4
pillow==12.0.0
texttable==1.7.0
